moved typed test cases into one file per test and removed redundant map type test
